:rotating_light: Remove some warnings

+++ plugins/processing/classification/src/algorithms/ovpCAlgorithmClassifierLDA.cpp
@@ -152,7 +152,7 @@ bool CAlgorithmClassifierLDA::train(const IFeatureVectorSet& dataset)
 	// Get class labels
 	for (size_t i = 0; i < m_nClasses; ++i)
 	{
-		m_labels.push_back(i);
+		m_labels.push_back(double(i));
 		m_discriminantFunctions.push_back(CAlgorithmLDADiscriminantFunction());
 	}
 
@@ -246,7 +246,7 @@ bool CAlgorithmClassifierLDA::train(const IFeatureVectorSet& dataset)
 	//We send the bias and the weight of each class to ComputationHelper
 	for (size_t i = 0; i < getClassCount(); ++i)
 	{
-		const double examplesInClass = nClasses[i];
+		const double examplesInClass = double(nClasses[i]);
 		if (examplesInClass > 0)
 		{
 			const size_t totalExamples = dataset.getFeatureVectorCount();
