+++ toolkit/src/algorithms/classification/ovtkCVector.hpp
@@ -5,6 +5,7 @@
 
 namespace OpenViBE {
 namespace Toolkit {
+	
 template <class TParent>
 class TVector : public TParent
 {
@@ -12,9 +13,9 @@ public:
 
 	explicit TVector(IMatrix& matrix) : m_matrix(matrix) { }
 
-	uint32_t getSize() const override { return m_matrix.getBufferElementCount(); }
+	size_t getSize() const override { return m_matrix.getBufferElementCount(); }
 
-	bool setSize(const uint32_t size) override
+	bool setSize(const size_t size) override
 	{
 		m_matrix.setDimensionCount(1);
 		m_matrix.setDimensionSize(0, size);
@@ -23,9 +24,9 @@ public:
 
 	double* getBuffer() override { return m_matrix.getBuffer(); }
 	const double* getBuffer() const override { return m_matrix.getBuffer(); }
-	const char* getElementLabel(const uint32_t index) const override { return m_matrix.getDimensionLabel(0, index); }
+	const char* getElementLabel(const size_t index) const override { return m_matrix.getDimensionLabel(0, index); }
 
-	bool setElementLabel(const uint32_t index, const char* label) override
+	bool setElementLabel(const size_t index, const char* label) override
 	{
 		m_matrix.setDimensionLabel(0, index, label);
 		return true;
